provide conversion of NvAttribtes to FFI type
The SAPI takes TPMA_NV while we are representing it as NvAttributes.
Before this change you could go only in one direction and now it is
bi-directional. Provide a simple way to create a NvAttribute to be
converted to a TPMA_NV by using derive_builder to provide the builder
pattern over the various booleans.